+++ src/app/(non-navbar)/email-signup/_component/EmailAuth.tsx
@@ -1,13 +1,15 @@
 import Button from "@/app/_component/common/atom/Button";
 import SigninInput from "@/app/(non-navbar)/email-signin/_component/SigninInput";
+import SignupEmailInfo from "./SignupEmailInfo";
+import SignupEmailWarning from "./SignupEmailWarning";
 
 interface Props {
   setStep: React.Dispatch<React.SetStateAction<number>>;
 }
 
-const SignupStepTwo = ({ setStep }: Props) => {
+const EmailAuth = ({ setStep }: Props) => {
   const handleNextStep = () => {
-    setStep(3);
+    setStep(2);
   };
 
   return (
@@ -15,16 +17,24 @@ const SignupStepTwo = ({ setStep }: Props) => {
       <SigninInput
         id="email-auth"
         name="email-auth"
-        title="비밀번호 (영문+숫자+특수문자 8~20자리)"
+        title="이메일"
         type="email"
+        theme="button"
       />
-      <SigninInput
-        id="certification"
-        name="certification"
-        title="비밀번호 확인"
-        type="number"
-      />
+      <div className="w-full -mt-1">
+        <SignupEmailWarning />
+      </div>
+      <div className="flex justify-center w-full mt-9">
+        <SigninInput
+          id="certification"
+          name="certification"
+          title="인증번호"
+          type="number"
+          theme="count"
+        />
+      </div>
       <div className="flex flex-col justify-end w-full grow mb-3">
+        <SignupEmailInfo />
         <div className="mt-6">
           <Button text="다음" theme="wide" onClickFn={handleNextStep} />
         </div>
@@ -33,4 +43,4 @@ const SignupStepTwo = ({ setStep }: Props) => {
   );
 };
 
-export default SignupStepTwo;
+export default EmailAuth;
